Tweak logo colors and positioning
The "G" of "GCW" aligns with the text inside the menu labels now,
instead of with the cursor edge.

The slightly larger logo font is avoid a particularly blurry upper
edge.

The red color logo was tweaked to better match the logo bar background,
while still matching the main menu cursor as well (as in, looking
similar: it is no longer exactly the same color).